+++ src/aliases.js
@@ -1,3 +1,3 @@
-THREE.Bootstrap.registerAlias('empty', ['size', 'fill', 'loop', 'time']);
+THREE.Bootstrap.registerAlias('empty', ['renderer', 'bind', 'size', 'fill', 'loop', 'time']);
 THREE.Bootstrap.registerAlias('core', ['empty', 'scene', 'camera', 'render']);