+++ src/extra/controls.js
@@ -13,7 +13,7 @@ THREE.Bootstrap.registerPlugin('controls', {
     three.controls = null;
 
     this._camera = three.camera || new THREE.PerspectiveCamera();
-    this.change({}, three);
+    this.change(null, three);
   },
 
   uninstall: function (three) {
@@ -22,9 +22,8 @@ THREE.Bootstrap.registerPlugin('controls', {
 
   change: function (event, three) {
     if (this.options.klass) {
-      if (this.klass !== this.options.klass) {
+      if (!event || event.changes.klass) {
         three.controls = new this.options.klass(this._camera, three.renderer.domElement);
-        this.klass = this.options.klass;
       }
 
       _.extend(three.controls, this.options.parameters);