+++ CHANGES.md
@@ -1,3 +1,4 @@
+* Move 'three.bind' and 'three.renderer' into their own plugins
 * Ad-hoc aliases to override plug-ins on init
 * Diff changes made with api.set(), fire 'change' only on change.